# Conflicts:
#	src/main/java/ru/example/group/main/aspect/LoggerAspect.java
#	src/main/java/ru/example/group/main/controller/UserRegisterController.java
#	src/test/java/ru/example/group/main/MainTest.java

+++ src/main/java/ru/example/group/main/repository/JwtBlacklistRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.example.group.main.entity.JwtBlacklistEntity;
 
 @Repository
-public interface JwtBlacklistRepository extends JpaRepository<JwtBlacklistEntity, Integer> {
+public interface JwtBlacklistRepository extends JpaRepository<JwtBlacklistEntity, Long> {
     JwtBlacklistEntity findJwtBlacklistEntityByJwtBlacklistedToken(String token);
 }
