+++ src/main/java/ru/example/group/main/repository/SocialNetUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import ru.example.group.main.entity.UserEntity;
 
 @Repository
-public interface SocialNetUserRepository extends JpaRepository<UserEntity, Integer> {
+public interface SocialNetUserRepository extends JpaRepository<UserEntity, Long> {
 
     UserEntity findUserEntityByEmail(String email);
 }
